Add discord OAuth support
Adds support for Discord as an OAuth provider.
Fetches user info and also the guilds (servers) that user is a member
of, and returns the IDs of said servers as the Groups userinfo
attribute. This allows, in conjunction with userfile, to authenticate
users based on their membership of discord servers.